requirements.txt
click==8.1.3
cycler==0.11.0
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
scipy==1.7.2
seaborn==0.11.2
six==1.16.0
tqdm==4.64.0
urllib3==1.26.12
Werkzeug==2.2.2
